refactor(yolo26-sem): move semantic-seg class-name helpers to dedicated module
validate_class_names and resolve_background_class_id are semantic-segmentation
specific and don't belong in the generic model_packages parsing module. Move
them to common/roboflow/semantic_segmentation.py (shared by the DeepLabV3+ and
YOLO26-sem backends) and repoint all six backend imports. Tests move to
test_semantic_segmentation.py.
