+++ inference_models/inference_models/models/deep_lab_v3_plus/deep_lab_v3_plus_segmentation_torch.py
@@ -23,6 +23,8 @@ from inference_models.models.common.roboflow.model_packages import (
     ResizeMode,
     parse_class_names_file,
     parse_inference_config,
+)
+from inference_models.models.common.roboflow.semantic_segmentation import (
     resolve_background_class_id,
 )
 from inference_models.models.common.roboflow.post_processing import (
