+++ inference_models/inference_models/models/yolo26/yolo26_semantic_segmentation_torch_script.py
@@ -22,6 +22,8 @@ from inference_models.models.common.roboflow.model_packages import (
     ResizeMode,
     parse_class_names_file,
     parse_inference_config,
+)
+from inference_models.models.common.roboflow.semantic_segmentation import (
     resolve_background_class_id,
     validate_class_names,
 )
