+++ inference_models/tests/unit_tests/models/common/roboflow/test_model_packages.py
@@ -20,8 +20,6 @@ from inference_models.models.common.roboflow.model_packages import (
     parse_inference_config,
     parse_key_points_metadata,
     parse_trt_config,
-    resolve_background_class_id,
-    validate_class_names,
 )
 
 
@@ -270,42 +268,6 @@ def test_parse_class_names_file_when_valid_config_provided(
 
     # then
     assert result == ["some", "other", "yet-another"]
-
-
-def test_resolve_background_class_id_when_background_present() -> None:
-    # when
-    result = resolve_background_class_id(["background", "road", "sidewalk"])
-
-    # then
-    assert result == 0
-
-
-def test_resolve_background_class_id_is_case_insensitive() -> None:
-    # when
-    result = resolve_background_class_id(["road", "Background", "sidewalk"])
-
-    # then
-    assert result == 1
-
-
-def test_validate_class_names_with_minimal_binary_class_list() -> None:
-    # background + a single foreground class is the minimal valid package
-    validate_class_names(["background", "object"])  # does not raise
-
-
-def test_validate_class_names_with_multiclass() -> None:
-    validate_class_names(["background", "road", "sidewalk"])  # does not raise
-
-
-def test_validate_class_names_when_background_absent() -> None:
-    with pytest.raises(CorruptedModelPackageError):
-        validate_class_names(["road", "sidewalk", "building"])
-
-
-def test_validate_class_names_when_no_foreground_class() -> None:
-    # background present but no foreground class -> invalid semantic-seg package
-    with pytest.raises(CorruptedModelPackageError):
-        validate_class_names(["background"])
 
 
 def test_parse_inference_config_when_path_does_not_exists() -> None:
